feat(HMS-1135): improve arch compatibility
Previously:
The InstanceTypesSelect would only show results for an architecture if
the architecture returned by the backend was exactly matching the one
asked by the user.

This proved to be problematic because some architecures can have many
names. For instance `aarch64` can be interchangeably used with `arm64`.

To to make the ProvisionningWizard more flexible, the instance selection
dropdown now is using the same name mapping for the architectures as the
one found in the backend. See:
https://github.com/RHEnVision/provisioning-backend/blob/0f3abc7328ebda3314d7496544b019a36a061152/internal/clients/architecture.go#L25

Refs HMS-1135